--- src/main/java/com/gb/market/market/dtos/OrderInfo.java
@@ -1,9 +1,0 @@
-package com.gb.market.market.dtos;
-
-import lombok.Data;
-
-@Data
-public class OrderInfo {
-    private String address;
-    private String phone;
-}
